Fix null error with mdAutocomplete
Currently programmatically changing the value of the autocomplete to null or undefined throws an error because the onInput function still runs which opens the panel and then attempts to check if it has changed which throws an error. Proposed change is to check if the search term is null or undefined stop the onInput function from continuing.